fix native closure in non-threadsafe mode
