+++ src/main.rs
@@ -24,6 +24,7 @@ mod io;
 mod lexer;
 mod libraries;
 mod macros;
+mod native_closure;
 mod parser;
 mod sourcecode;
 mod symbol;
